backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.7
redis==5.0.0
chromadb==0.4.22
langchain==0.1.11
langchain-community==0.0.28
transformers==4.38.2
torch==2.2.1
sentence-transformers==2.5.1
python-magic==0.4.27
pydantic-settings==2.2.1
python-decouple==3.8
celery==5.3.6
openai-whisper==20231117
opencv-python==4.9.0.80
pypdf==4.1.0
loguru==0.7.2
PyPDF2
llama-cpp-python
